removed pipe and redir completely, will redo based on current format
