+++ minishell.h
@@ -12,18 +12,22 @@
 
 #ifndef MINISHELL_H
 # define MINISHELL_H
-#include "./libft/libft.h"
-#include "./printf/includes/ft_printf.h"
-#include <unistd.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <signal.h>
-#include <stdbool.h>
-#include <errno.h>
-#include <sys/types.h>
-#include <sys/wait.h>
-#include <readline/readline.h>
-#include <readline/history.h>
+# include "./libft/libft.h"
+# include "./printf/includes/ft_printf.h"
+# include <unistd.h>
+# include <stdio.h>
+# include <stdlib.h>
+# include <stddef.h>
+# include <signal.h>
+# include <limits.h>
+# include <stdbool.h>
+# include <errno.h>
+# include <fcntl.h>
+# include <sys/types.h>
+# include <sys/wait.h>
+# include <sys/stat.h>
+# include <readline/readline.h>
+# include <readline/history.h>
 
 //Macros
 # define showCmds "Custom commands: echo | env | exit | showpath\nSystem commands: non-argument syscalls might work e.g. ls | pwd\n"
@@ -37,6 +41,10 @@
 # define SUCCESS 0
 # define FAILURE 1
 
+# define STDIN 0
+# define STDOUT 1
+# define STDERR 2
+
 typedef struct s_data
 {
 	char	*cwd;
@@ -45,8 +53,6 @@ typedef struct s_data
 	char	**args;
 	char	*input;
 	char	*cmd;
-	char	*pipe_fd;
-	char	*pipe_out;
 	pid_t	pid;
     struct  s_cmd   *cmds;
 	struct s_data	*next;
@@ -58,24 +64,9 @@ typedef struct s_cmd
 	char	*command;
 	char	*path;
 	char	**args;
-	bool	pipe_out;
-	int		*pipe_fd;
-	struct s_io	*io_fd;
 	struct s_cmd	*next;
 	struct s_cmd	*prev;
 }	t_cmd;
-
-typedef struct s_io
-{
-	char	*infile;
-	char	*outfile;
-	char	*heredoc_delimiter;
-	bool	heredoc_quotes;
-	int		fd_in;
-	int		fd_out;
-	int		stdin_temp;
-	int		stdout_temp;
-}	t_io;
 
 //Initialisation
 void	initData(t_data *data, char **env);
@@ -96,27 +87,16 @@ bool	valid_env(char *var);
 int	env_index(char **env, char *var);
 char	*env_value(char **env, char *var);
 
+//Sys call
+void	ft_fork_and_exec(t_data *data, char *input, char **argv);
+void	sys_call(char *input, t_data *data);
+
 //Processing
 void	start(t_data *data, t_cmd *c);
 int	process(char *input, t_data *data);
 
 //Quotes
 char	*ft_quote(char *input, int i, int j);
-
-//Pipe and Redir
-bool	reinit_io(t_io *io);
-bool	redir_io(t_io *io);
-bool	check_io(t_io *io);
-void	init_io(t_cmd *c);
-void	close_pipes(t_cmd *c, t_cmd *c_temp);
-bool	malloc_pipes(t_data *data);
-bool	init_pipes(t_cmd *cmd, t_cmd *c);
-//Pipe and Redir (in main.c)
-void	close_fds(t_cmd *cmd, bool check);
-int	get_child(t_data *data, t_cmd *cmd);
-int	init_child(t_data *data, t_cmd *cmd);
-int	prep_exec(t_data *data, t_cmd *cmd);
-int	execute(t_data *data, t_cmd *cmd);
 
 //Builtins
 int	ft_cd(char *input, t_data *data);
